# static/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "The Cerberus is an users' data keeper. The Cerberus encrypts data and pushes it into S3.",
    "title": "Cerberus",
    "version": "1.0.0"
  },
  "basePath": "/v1",
  "paths": {
    "/pdv": {
      "post": {
        "security": [
          {
            "public_key": []
          },
          {
            "signature": []
          }
        ],
        "consumes": [
          "application/octet-stream"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Cerberus"
        ],
        "summary": "Encrypts and puts PDV data into storage.",
        "operationId": "SavePDV",
        "parameters": [
          {
            "description": "file's raw bytes",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "type": "string",
              "format": "binary"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "pdv was put into storage",
            "schema": {
              "$ref": "#/definitions/SavePDVResponse"
            }
          },
          "400": {
            "description": "bad request",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "401": {
            "description": "signature wasn't verified",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "internal server error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/pdv/{address}": {
      "get": {
        "security": [
          {
            "public_key": []
          },
          {
            "signature": []
          }
        ],
        "produces": [
          "application/octet-stream",
          "application/json"
        ],
        "tags": [
          "Cerberus"
        ],
        "summary": "Gets and decrypts PDV from storage.",
        "operationId": "ReceivePDV",
        "parameters": [
          {
            "type": "string",
            "description": "PDV's address",
            "name": "address",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "PDV from storage",
            "schema": {
              "type": "file"
            }
          },
          "400": {
            "description": "bad request",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "401": {
            "description": "signature wasn't verified",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "403": {
            "description": "access to file is denied",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "internal server error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/pdv/{address}/meta": {
      "get": {
        "tags": [
          "Cerberus"
        ],
        "summary": "Return PDVs meta.",
        "operationId": "GetPDVMeta",
        "parameters": [
          {
            "type": "string",
            "description": "PDV's address",
            "name": "address",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/definitions/PDVMeta"
          },
          "400": {
            "description": "bad request",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "401": {
            "description": "signature wasn't verified",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "404": {
            "description": "PDV doesn't exist"
          },
          "500": {
            "description": "internal server error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Error": {
      "type": "object",
      "properties": {
        "error": {
          "description": "error message",
          "type": "string",
          "x-go-name": "Error"
        }
      },
      "x-go-name": "error",
      "x-go-package": "github.com/Decentr-net/cerberus/internal/server"
    },
    "SavePDVResponse": {
      "type": "object",
      "properties": {
        "address": {
          "description": "Put file address",
          "type": "string",
          "x-go-name": "Address"
        }
      },
      "x-go-name": "savePDVResponse",
      "x-go-package": "github.com/Decentr-net/cerberus/internal/server"
    }
  },
  "securityDefinitions": {
    "public_key": {
      "description": "Blockchain account's public key",
      "type": "apiKey",
      "name": "Public-Key",
      "in": "header"
    },
    "signature": {
      "description": "Signature of request digest.\nDigest is sha256 sum of request: {body as is}+{request uri}.\nFor example:\n   Private key in hex: cfe43c70347c7e39084612d9448f3ed86ed733a33a67de35c7e335b3c4edc37d\n   Request url: http://localhost/v1/pdv\n   Body: {\"some\":\"file\"}\n   Digest will be made from `{\"some\":\"file\"}/v1/pdv`\n   Digest in hex:\n     4a1084d05820d60aee9ce600227ca2290ef63e80e5227215b58b023ec6876799\n   Signature in hex:\n     28eff4676d7839648dda925ba92d447dd7552e177a302f32681fc76278088f9f1fb98051666aa02dd80f7d9b7c01d42ea1abbb3e65de8f1fd04be7b747fb0692",
      "type": "apiKey",
      "name": "Signature",
      "in": "header"
    }
  },
  "security": [
    {
      "public_key": []
    },
    {
      "signature": []
    }
  ]
}